修复Mixin应用时机
避免Mixin无法找到class，同时修改Mixin内容，先确保可以渲染出正确字符，内容正确性延后处理
