+++ gradle.properties
@@ -93,7 +93,7 @@ mixinsPackage = mixins
 # Specify the core mod entry class if you use a core mod. This class must implement IFMLLoadingPlugin!
 # This parameter is for legacy compatibility only
 # Example value: (coreModClass = asm.FMLPlugin) + (modGroup = com.myname.qz_uilibb) -> com.myname.qz_uilibb.asm.FMLPlugin
-coreModClass = mixins.EarlyMixins
+coreModClass = mixins.early.EarlyMixins
 
 # If your project is only a consolidation of mixins or a core mod and does NOT contain a 'normal' mod ( = some class
 # that is annotated with @Mod) you want this to be true. When in doubt: leave it on false!
