+++ src/main/java/club/heiqi/qz_uilib/mixins/late/MixinSmallFontRenderer.java
@@ -1,0 +1,20 @@
+package club.heiqi.qz_uilib.mixins.late;
+
+import club.heiqi.qz_uilib.Config;
+import club.heiqi.qz_uilib.fontsystem.impl.ReplaceFontRender;
+import mantle.client.SmallFontRenderer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(SmallFontRenderer.class)
+public class MixinSmallFontRenderer {
+
+    @Inject(method = "renderString*", at = @At("HEAD"), cancellable = true, remap = false)
+    public void renderString(String text, int x, int y, int color, boolean darkDropShadow, CallbackInfoReturnable<Integer> cir) {
+        if (Config.replaceOrigin) {
+            cir.setReturnValue(ReplaceFontRender.getInstance().renderString(text, x, y, color, darkDropShadow));
+        }
+    }
+}
