+++ src/main/resources/mixins.qz_uilib.late.json
@@ -1,0 +1,8 @@
+{
+  "required": false,
+  "minVersion": "0.8.5-GTNH",
+  "package": "club.heiqi.qz_uilib.mixins.late",
+  "refmap": "mixins.qz_uilib.refmap.json",
+  "target": "@env(DEFAULT)",
+  "compatibilityLevel": "JAVA_8"
+}
